+++ flask_sqlalchemy/model.py
@@ -3,6 +3,7 @@ import re
 import sqlalchemy as sa
 from sqlalchemy import inspect
 from sqlalchemy.ext.declarative import DeclarativeMeta, declared_attr
+from sqlalchemy.schema import _get_table_key
 
 from ._compat import to_str
 
@@ -72,6 +73,15 @@ class NameMetaMixin(object):
         If no primary key is found, that indicates single-table inheritance,
         so no table will be created and ``__tablename__`` will be unset.
         """
+        # check if a table with this name already exists
+        # allows reflected tables to be applied to model by name
+        key = _get_table_key(args[0], kwargs.get('schema'))
+
+        if key in cls.metadata.tables:
+            return sa.Table(*args, **kwargs)
+
+        # if a primary key or constraint is found, create a table for
+        # joined-table inheritance
         for arg in args:
             if (
                 (isinstance(arg, sa.Column) and arg.primary_key)
@@ -79,6 +89,15 @@ class NameMetaMixin(object):
             ):
                 return sa.Table(*args, **kwargs)
 
+        # if no base classes define a table, return one
+        # ensures the correct error shows up when missing a primary key
+        for base in cls.__mro__[1:-1]:
+            if '__table__' in base.__dict__:
+                break
+        else:
+            return sa.Table(*args, **kwargs)
+
+        # single-table inheritance, use the parent tablename
         if '__tablename__' in cls.__dict__:
             del cls.__tablename__