+++ tests/test_table_name.py
@@ -1,5 +1,7 @@
 import inspect
 
+import pytest
+from sqlalchemy.exc import ArgumentError
 from sqlalchemy.ext.declarative import declared_attr
 
 
@@ -191,3 +193,23 @@ def test_no_access_to_class_property(db):
             assert False
 
     assert ns.accessed
+
+
+def test_metadata_has_table(db):
+    user = db.Table(
+        'user',
+        db.Column('id', db.Integer, primary_key=True),
+    )
+
+    class User(db.Model):
+        pass
+
+    assert User.__table__ is user
+
+
+def test_correct_error_for_no_primary_key(db):
+    with pytest.raises(ArgumentError) as info:
+        class User(db.Model):
+            pass
+
+    assert 'could not assemble any primary key' in str(info.value)